backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
pg8000>=1.31.2
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1
spacy==3.7.5
faker==25.8.0
pytest==8.2.2
httpx==0.28.1
python-multipart==0.0.9
PyPDF2>=3.0.0
python-docx>=1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
fastapi-mail==1.4.1
twilio>=9.0.0
